Move files around for proper setup (no fixes yet)
