+++ src/covid19sim/base.py
