+++ src/covid19sim/frozen/utils.py
