+++ src/covid19sim/models/log_risk_mapping.npy
