+++ src/covid19sim/toy.py
