+++ src/covid19sim/utils.py
